requirements.txt
fastapi[standard]
uvicorn
sqlalchemy
jinja2
passlib[bcrypt]
python-jose
itsdangerous
pytest